Create HARD RESET
When you run this script, it introduces "Dumbledore, the Wizard of Apple Device Resets," and simulates initiating a hard reset on an Apple device running iOS 16 or higher. It conjures a magical incantation to perform the reset, adding a playful twist to the technical process. Feel free to adjust the phrases or add more magical elements to make it even more whimsical!